a11y: Drop unintended check for window type being non-zero
Commit 789a2310c0d09cb15cc9ecd793843a5abb3586c8
("Remove Qt::WindowType::Desktop in QAccessbileObject")
presumably was supposed to remove the check for the
window type not being Qt::Desktop altogether, not
changing it to check that the the type is non-zero instead.

While at it, also update the comment to no longer
mention the desktop.

This amends 789a2310c0d09cb15cc9ecd793843a5abb3586c8.

Task-number: QTBUG-140514
Change-Id: I53527db83ac67d343567cdeaba8204a95195c20b
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>