Expand QRhiWidget color buffer size docs
Traditionally we always talk about the "supersampling" example
with textureSize type of properties (following Quick's docs from
the Qt 5.0 days), whereas in reality the more common use these
days is to set a smaller size so that one gets content upscaled
from a lower resolution. (e.g. because the developer decides
that rendering at 1080p is sufficient with a fullscreen QRhiWidget
on a 4K screen) Expand also on the consequences of a > 1 device
pixel ratio, to make it clearer that it is in effect an implicit
fixedColorBufferSize = 2 * pixel_size, when it comes to the
visual results.

Pick-to: 6.10 6.9 6.8
Change-Id: I9ede72ddcc314798fa2b17115cca288d7d90da98
Reviewed-by: Andy Nichols <andy.nichols@qt.io>