QWinRegistryKey: disable moves and swap()
Since 589bfddc7c44a92e7e3f98a1e73871793c153ab7, the class is-a QObject
and has a new data member, m_keyChangedEvent. Neither swap() nor the
move SMFs handle the new member, and it's unclear how they should
handle it, seeing it is tied to both m_key and the QObject identity of
its enclosing class, the latter of which cannot be moved or swapped.

De-QObject-ify'ing the class again (QTBUG-140671) would allow to
define the semantics of subsribing to valueChanged() notifications
vis-a-vis move semantics. Until then, and as a hot-fix, disable the
move SMFs and swap().

Task-number: QTBUG-140725
Pick-to: 6.10
Change-Id: I9cecbe2afb98e5caa6889c2591fdd6a59106dab3
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>