QWinRegistryKey: de-pessimize stringValue()
Don't create a QString object that we might not need (if the optional
already contains one).

Using value_or() with an argument that is not of trivial type is like
having a function default argument of the same type: it's constructed
whether we need it or not; one should use overloading instead
(QTBUG-98117).

Rewrite to the equally-readable check-and-return-else-default-construct.

Amends 40523b68c14bf618bdc2d5438deebf34627be3af.

Pick-to: 6.10 6.8
Task-number: QTBUG-98117
Change-Id: Iecd7fa9c3705c8416b5609e6d157079e10bfca59
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>