Decouple tst_Moc::warnOnMultipleInheritance from include paths
The test was using QWindow as one of the QObject based classes, but
didn't pass an exhaustive list of include paths to the moc command
invocation, like we do when moc is invoked from the build system.

As a result, it failed to find the qwindow.h header when included
as <QtGui/qwindow.h>, which moc surprisingly does not treat as an
error or even a warning.

Which in turn lead moc to not realize that the leaf class was
picking up QObject from both base classes, missing the expected
warning of the test.

The test is now fully self-contained, with both QObject base
classes in the test file. We don't even have to include anything
else, but for good measure we include <QtCore/qobject.h>.

Change-Id: I3aed4dddfdfe73a4d814ebefefeb282dab361dcd
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>