Update contrastPreference correctly on Android
Android has an option called "Color contrast" in the
Settings app, found under "Display and Touch".

Qt has a new property in 6.10 called contrastPreference, which can
either be `NoPreference` or `HighContrast`. The intention is to use
native API's to determine if the user has enabled a setting on the
platform level, to increase the overall contrast on the system.
If that is the case, the contrastPreference property should return
`HighContrast`.

This property is then used by our built-in styles, to add more outlines
to controls and potentially change some of the palette colors that are
in use. It can also be used by custom styles.

We're already detecting if the system is running in high contrast mode
on Windows, macOS, Gnome, and WebAssembly. Android took a bit longer to
implement due to the ambiguity of their API. On Android the user has
the option to set the "Color contrast" to either `Default`, `Medium`
or `High`. For now, the contrastPreference property will only return
`HighContrast` if the "Color contrast" is set to `High` only.

When changing this setting on Android, the Activity is recreated.
It is desireable to handle this configuration change ourselves, however,
we've not been able to find a good approach here. Most configuration
changes can be added to a list in AndroidManifest.xml to override the
default behavior of resetting the Activity, but 'colorMode' or any other
handler doesn't seem to work when changing the contrast setting.

Fixes: QTBUG-139294
Pick-to: 6.10
Change-Id: Iaebb89c79feb749655132d2db83c1bb1a67b301d
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>