QRM: correctly forward a forwarding reference in updateTarget
Even though we know that we always call updateTarget with rvalue
reference as the copy (we don't need the copy anymore), code checker
correctly complains about the std::move of a forwarding reference.

Fix that, and generalize the code further; since we have an rvalue
reference on the right hand side of the assignment operation, we can
always try to (move-)assign, even if the types are different. Only try
swap if we can't assign.

Amends 21eee24a78e76a8763f81632b1e4368846b64ab3.

Change-Id: Ib0305db5b08b634174f388497ab47d05fb8c1838
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>