qdatetime: reject 2 digit years as invalid in rfc2822 format
Reject all digit year notation in rfcDateImpl that don't use 4 digits,
and return an invalid date.
Using 2 digit year is marked as obsolete in RFC2822, and the previous
implementation interpreted 2 digit years as is, like 22 became 22 (not
1922, not 2022, but 22) for example.

Also note that the documentation already explicitly says that the year
should be in the form of "yyyy" and therefore contain 4 digits.

This helps fix a bug in qtdeclarative's date parser in
qv4dateobject.cpp, where we only apply our special fallbacks (for
example for 2 digit years) when the QDateTime parsing methods return
an invalid QDateTime.

Pick-to: 6.10 6.8
Task-number: QTBUG-100377
Change-Id: I12c61ca35f38797e954e4126763252026ecbb008
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>