Mark QWinRegistryKey as security-critical
It's a data parser, even though the parsing is generally rather
trivial, but it's also low-level, with unsecure C APIs being
called. The data is external, because the Windows registry is open
like the file system is. So if parsing from the file system is
critical (cf. QSettings), then so is parsing from the registry.

The header only contains trivial implementation, so mark it as
significant.

Task-number: QTBUG-135190
Pick-to: 6.10 6.8
Change-Id: I44ad39ec12460e862f97911bc68166cb2f7122cf
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>