Mark qcborcommon_p.h as security-critical
This might be a bit controversial, but this header configures the API
of TinyCBor for use in Qt, so it's not just declarations. TinyCBor, as
third-party component, is listed as security-critical in its
qt_attribution.json, so I opted to mark the glue code as critial, too.

Amends cd4813a3c304a6d9d5d8096d91a8fc297818d624.

QUIP: 23
Task-number: QTBUG-135194
Pick-to: 6.10 6.8
Change-Id: I0042a8677eff716eb1924dfc211c740d164029cd
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>