Mark remaining files in src/corelib/serialization as security-insignificant
The .gitignore file is obviously insignificant. The file format allows
comments, so mark it. Invent a new reason string for that.

The shell script is also insignificant. While it runs qlalr for the
security-critical QtXmlStream classes, as a build-tool, it's not
security-critical itself, and, unlike e.g. util/normalize/main.cpp, it
doesn't itself contain code that ends up being compiled. Invent a new
reason string for that.

This completes the review of src/corelib/serialization:

  $ for i in $(find src/corelib/serialization -type f); do
      if ! grep -qE '^ *(//|#) *Qt[ -]Security +score:' "$i"; then
          echo "$i"
      fi
  done
  <nothing>

QUIP: 23
Fixes: QTBUG-135194
Pick-to: 6.10 6.8
Change-Id: Id5d18244fe0e9d18b8891500a3a946ac530671a4
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>