tst_qtjson: Increase stack size when run with ASan
Amends 1756a84756807a9849aa507e77845dfdf31b8020, which added the
"nesting" test. That has 1024 nested arrays, which should we should stop
parsing on. Under normal circumstances, we do that just fine. But with
ASan, the size of each stack frame is pretty large and we end up
overflowing the 1 MB stack on Windows.

So increase to 8 MB, matching Unix platforms.

Fixes: QTBUG-140452
Pick-to: 6.10 6.8
Change-Id: I28de3084bba4b9216b5ffffdcbd69278ae663b5b
Reviewed-by: Piotr Wierciński <piotr.wiercinski@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>