QAbstractSocket/QHostAddress: ensure definition of static constexprs
Amends commit 969337bcfd6af6d91b988e4b412703274a0b5877, which introduced
these constants. We've since learned that those either need an
out-of-line definition or need to be in a non-exported base[1]. This
commit uses both strategies: an out-of-line definition for the symbols
that are used in Qt 6 and thus have an ABI requirement to stay the same,
and moving to a base class for the Qt 7 ones, which don't yet.

[1] https://lists.qt-project.org/pipermail/development/2024-January/044888.html

Pick-to: 6.10 6.8
Change-Id: Id6fb9fd431ee59534769fffd916b9db68c4e18a9
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>