Q{Explicitly,}SharedDataPointer: share most code
The majority of the code for the two classes is identical, so we can
share code. QExplicitlySharedDataPointerV2's is different... and wrong?

We need explicit constructors, assignment operators, and destructors to
prevent early instantiation of the base class's equivalents, which break
compilation of classes using the QT_DECLARE_Q(E)SDP_SPECIALIZATION_DTOR
macros.

Change-Id: I9e3cde075b537c61e8fcfffdb7a1a2d74b399967
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>