QTimeZone: remove remnant of QSharedDataPointer<QTimeZonePrivate>
Amends commit ae6186c7e8cfdb9420b9119f5affbba7d069598d, which replaced
the implicit d-pointer using QSharedDataPointer of the original Qt 5.2
design with an explicit implementation, with support for a "short
object" of just an offset.

There's a QExplicitlySharedDataPointer<QTimeZonePrivate> in
qtimezone.cpp, which currently doesn't detach at all. If and when it
does, reimplementing this may be needed.

Pick-to: 6.10 6.8
Change-Id: I92b4b30f60689d0a1f5ffffd53f6f71cd1834f15
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>