Android: improve setting of platform window states and flags
The current implementation had two issues:

1. Setting a flag before the window is shown has no effect, becuase
   the android platform window had its own flags data, but before
   the window is shown, the platform window isn't yet created, so
   that initial flag setting is lost.
2. The system UI was never updated when a flag has changed, as in
   the case of Qt::ExpandedClientAreaHint which affects the system
   the system UI visibility.

This patch gets rid of rid of the platform window's separate
internal tracking of states and flags to make things simpler.
Also, it changes the way updateSystemUiVisibility() works by
passing states/flags to it to be more flexible to handle all
changes and cases.

Fixes: QTBUG-140830
Pick-to: 6.10 6.8 6.5
Change-Id: Iff9f1b02a0a71320a3514e462de0cd6c1b8589fd
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>