Fix text offset when drawing a QPicture
When replaying text in a QPicture, it would try to detect the baseline
of the text by taking the ascent of the selected font. In most cases
this will work, but if the text contains characters from other fonts
(due to font merging) then we need to take the ascents of these fonts
into account as well, since the actual distance to the baseline
will be the maximum ascent of all fonts.

To enable this, we implement support for Qt::AlignBaseline in
qt_format_text() so that it can apply the correct offset by actually
looking at the ascent of the first text line.

There is a complication though: Since Qt::TextSingleLine and
Qt::AlignBaseline share the same value, they cannot be combined
in a single flags field. So we expand the qt_format_text() to
allow passing text flags and alignment separately. To avoid
accidentally enabling AlignBaseline for normal drawText() calls,
we mask this out when passing the alignment in.

The code has been this way since 2007, so applications may have
worked around this bug for years. Therefore, we only enable the
changed behavior for the latest QPicture format. By setting an
older format you can still get the old behavior.

[ChangeLog][QtGui] Fixed an issue where text drawn into a QPicture
might be offset vertically when replayed.

Fixes: QTBUG-139243
Change-Id: I5fe7edc561bd5f666c8c64fb2ed603011ddded47
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>