CMake: Call source_group("Resources") for qt_add_resources files
Qt Creator parses the cmake file api response to group a target's
source files into a project tree. CMake has a hardcoded list of file
extensions that it considers as resources, which it places into
a "Resources" source group.
This list is not exhaustive, which causes for example png files to be
under Resources, but not gif files.

Record the list of files added by qt_add_resources into a target
property. When finalizing the target's source groups, place these
files into the "Resources" source group, except for qml and js files
which are kept as source files.

The source_group call respects nested resource file paths, so if a
resources/images/a.png is added, it will be placed under
Resources/resources/images/a.png in the project tree.

Add an opt out variable QT_NO_AUTO_RESOURCE_SOURCE_GROUPS to disable
this behavior if it doesn't fit the project.

Pick-to: 6.8 6.10
Fixes: QTBUG-137936
Change-Id: Ic12481be32714cdd95ca33a39e0fc1c802505abd
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>