qt-testrunner: if a test logs a qFatal message, then treat it as crashed
QTest can, depending on the platform, clean-up properly after a qFatal.
In that case the test <Incident> is marked simply as FAIL, no further
tests are run, and the XML logfile is closed properly.
The only way to figure out that something went bad and the testrun was
aborted, is the qFatal message.

This is needed primarily for qt-testrunner to *not* consider re-running
individual failed test functions/datatags, to avoid missing the ones
that did not start at all.

Fixes: QTQAINFRA-7349
Fixes: QTQAINFRA-7378
Pick-to: 6.11 6.10 6.8
Change-Id: I77276f37e0e6668048c78563560af0e15101be0e
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>