Add missing nullptr guard to QTextStreamPrivate::setupDevice(QIODevice*)
Amends commit e3c290e1947515992821e6bf97d74d65c9254271, which missed
this. The guard is necessary to prevent a connection attempt with a
nullptr when QTextStream::setDevice(QIODevice*) is called with a nullptr
as a means to dissociate it from its current device without attaching it
to a new one.

Pick-to: 6.10
Change-Id: I51d417fc9d3549b60c1f9e3e5b35e26b2d8fb1a9
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>