appendNodeText: Don't split emojis with surrogate chars in two
When parsing a document that assigns an anchor to text, we assign
this anchor to the next immediate character. If this character was the
low surrogate in a pair, we would end up splitting the surrogate and
messing up the text. We use QStringIterator to make sure we always
process surrogate pairs together and add the full pair to the resulting
text.

Fixes: QTBUG-140929
Pick-to: 6.10
Change-Id: I63ee03c251004d1a138f97462723fcc3798b9147
Reviewed-by: Albert Astals Cid <aacid@kde.org>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>