QToolButton: don't override user-set popup mode
The attempt to unify the popup mode for a QAction with a QMenu resulted
in the ignorance of the popup mode set by the user. Therefore add some
logic to know if the mode can be set to the 'new' default mode when the
QAction has a corresponding QMenu.
This amends 0f6135ae6d97f77c1a737a26f67bf3242940e11c

Pick-to: 6.10
Task-number: QTBUG-140148
Fixes: QTBUG-140793
Change-Id: Ie093744f1474dbd8c00112345e77497cc3177b2c
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>