Fix QLineEdit frame color when using gradient style sheet and Fusion
When using a  gradient style QBrush via style sheet, an
invalid color is set on the widget palette for the Window role.

This color is used to in QFusionStylePrivate::outline() to calculate
the outline color, resulting in a black/invalid color as well for
PE_FrameLineEdit. Fall back to the widget palette in case an invalid
color is encountered.

Pick-to: 6.10
Fixes: QTBUG-138960
Change-Id: Id1b5320476e59a743dfd0cf489032fe34e47d422
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>