wayland: Fix subsurface buffers outliving the backing store queue
In the path to handle subsurfaces that are painted as part of another
window, QWaylandShmBacking store may create temporary buffers for the
subsurface window.

This buffer has it's own lifespan managed by autodeletion from the
compositor not from the backing store. As it can outlive the
QWaylandShmBackingStore it should not share the same event queue as
having an object outlive the event queue can be problematic. As we
create a new buffer every time the custom event queue is not needed
here.

Amends: a0161a2b783c1cbea8f1faaa6fe9f14160515900

Pick-to: 6.10
Change-Id: I7815cf8c9fc539e739499a9e637b1e52cb9c92f3
Reviewed-by: Vlad Zahorodnii <vlad.zahorodnii@kde.org>