QMdiSubWindow: Respect null icon setting on Mac
On Mac, a QMdiSubWindow title bar by default shows no icon. However,
if an icon is then set on it, there would be no way to return to the
initial no-icon-shown state. Setting a null icon would only result in
the default window icon to be shown instead. This commit fixes that.

Fixes: QTBUG-139345
Pick-to: 6.10 6.8
Change-Id: I456a2cbdc4b81312a2959bc0f84a1dd545908b1f
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>