Teach QKeySequence::toString(Native) about C0 Control Pictures
The Unicode C0 control codes correspond to the 32 control characters
in ASCII from 0x00 and 0x1f. Since these are unprintable in their
basic form, it makes sense to represent them as something printable
via toString(). The Unicode Control Pictures block provides exactly
that.

We skip 0x0 (Null) as we don't distinguish between a missing key
and the null character in Qt.

Although there are also Control Pictures for space (0x20), del
(0x7f), and others, we ignore these as they are not in C0.

[ChangeLog][QtCore] QKeySequence::toString() now maps the C0
control characters to their equivialent in the Unicode Control
Pictures block, if using QKeySequence::SequenceFormat::NativeText.

Change-Id: I4525d08478ebeb2a321b035c96029a22fa5ed749
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>