QGraphicsView tests: fix unittests for 'windows' style
QGraphicsView::mapFromScene() returns the coordinates in viewport
coordinates, the tests use them to trigger events in the view. Since the
view might contain a frame around the viewport, the coordinates differ
depending on the value of PM_DefaultFrameWidth. Some tests already had a
fuzzing compare to not fail for a frame width of 1 but this does not
work for a larger value.
Some tests also set a fixed size for the view where the frame width was
not considered which lead to unwanted scrollbars.

Fixes: QTBUG-141079
Change-Id: If09898e08d5e5558825db461f02b6847ae90fa29
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>