Set QGuiApplicationPrivate::lastCursorPosition on Enter event
As soon as a window is shown, if it happens to be under the mouse
cursor, lastCursorPosition should be set correctly. The fact that this
was not done was masked in Qt Quick by the separate tracking of current
mouse position in QQuickDeliveryAgent. We need the tracking to be
reliable in QtGui so that Qt Quick does not need to do that.

Pick-to: 6.10
Task-number: QTBUG-136976
Change-Id: Ida397beeb335bc054fd8a005c7dc65f2b7e48ffa
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>