QUndoStack: fix two Coverity COPY_INSTEAD_OF_MOVE warnings
Coverity complained that, since the ActionState struct contains a
QString, a copy is expensive while a move is possible.

Make it so.

To avoid explicit std::move() far away from the end of the scope
(which impairs readability), use C++17 if-with-initializer to reduce
the scope of the newX ActionState objects. This is a drive-by, since
to move from them, we need to de-const the variables, anyway.

Amends 31b0dadb0f371fc94652782c28024f135a0b6f4b.

Pick-to: 6.10 6.8
Coverity-Id: 896353
Coverity-Id: 896351
Task-number: QTBUG-138567
Change-Id: Id365d9bd91773452d16634a3b862f3d23ad650b9
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Ali Kianian <ali.kianian@qt.io>