TestLib: macOS - unconditionally disable display sleep
The tests only disabled display powersave for displays when a
QApplication was running. It did not prevent display powersave for
QGuiApplication, nor did it prevent any system sleep if the qApp
singleton was created as part of the unit test of for app-less tests.

To be on the safe side, we should prevent the display (and system) from
going to sleep while any test is running.

Change-Id: Icd8ddfda8c9fb8d551548def27ff65776bc82486
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>