QCborValue: rewrite nextUtf32Character to avoid narrowing char32_t → char16_t
Ought to fix Clang 21 -Wcharacter-conversion in this function.

Instead of storing the initial *ptr in r.c (a char32_t) and then
narrowing it back to char16_t for the call to
QChar::surrogateToUcs4(), store it in a separate variable and
construct an R only in the return statement.

Also move the len = 1 return from the end of the function to the
(inverted) if (c < 0x80) near the start of the function.

This is less assmebler-esque, so might execute slower (though it's all
values, so the optimizer should have no problem rewirting this to the
old code, if that was faster), but is much easier to read, which is
not the sneezed at in a security-critical component.

Amends d4c7da9a07dc1434692fe08a61ba22c794574c4f.

Pick-to: 6.10 6.8 6.5
Change-Id: I1ce610fa80e3d874106c9c5b3a1d50bf99f39732
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>