QChar: avoid char{32→16}_t narrowing in foldCase(char16_t*, char16_t*)
... which triggers Clang 21 -Wcharacter-conversion.

Instead of using the char32_t ucs4 variable to temporarily store the
read from *ch, keep using *ch until we have determined whether it
needs to be decoded as a surrogate pair or can simply be promoted to
char32_t.

Since this touches most lines of the function, anyway, take the
opportunity to rename the first argument from 'ch' to 'cur', and change
pointer arithmetic to indexing operations.

Amends the start of the public history.

Pick-to: 6.10 6.8 6.5
Change-Id: I47a026fd509da8fa3dee4b6be30ac4432148d9c6
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>