QLockFile: refactor local helper into static member functions
To prepare for the Windows backend doing extra work.
We want to open the file with FILE_SHARE_DELETE so that
we can eventually create the file with the DELETE permission,
enabling us to delete it without closing the handle first.

Since we can't do that through QFile directly we need to open
the file descriptor first and then pass that to QFile.

Change-Id: I0baf0ee3089b489f3bd1e0259cc06195e9256647
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>