Avoid float rounding errors in QDashStroker
Calculations which should have a result of zero can lead to slightly
different results when using floating point values. This patch fixes a
case which could trigger an assert although all values passed to the
public API were harmless.

Credit to OSS-Fuzz for finding an SVG file which triggered this as issue
429123947. I derived the test from the values in the file.

Pick-to: 6.10 6.8
Change-Id: I06d6f0f0bf6beb758f54423e16e9c5653ed8edf1
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>