tst_QUdpSocket: update the whitelisting of destinationAddress() support
We had hardcoded Linux, Windows, and the BSDs (which includes macOS and
FreeBSD), which are all OSes I can directly test. Instead of duplicating
the checks everywhere and often not doing it properly, let's centralize
with a static constexpr variable and a correct comment.

In addition, it turns out that all platforms in our CI support receiving
the IPv4 destination address. Indeed, QNX[1][2] and VxWorks
documentations indicate they do have IP_RECVDSTADDR.

Though, it looks like QNX fails to convey the destination address to a
dual-stack IPv6 socket when the sender is IPv4. That's probably an OS
bug. On FreeBSD, which also uses IP_RECVDSTADDR for IPv4, the
ancillary data was present.

[1] https://www.qnx.com/developers/docs/6.5.0SP1.update/com.qnx.doc.neutrino_lib_ref/i/ip_proto.html
[2] https://www.qnx.com/developers/docs/7.0.0/#com.qnx.doc.neutrino.lib_ref/topic/i/ip_proto.html

Change-Id: I6ee58e3f896e07c9eaa9fffdac51d092901ea855
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>