qt_testrunner: fix execution of qt_mock_test.py under testing on windows
tst_qt_testrunner.py  invokes  qt_testrunner.py  which invokes  qt_mock_test.py.

On Windows a python script is not directly executable. It needs to be
executed as "python.exe script.py". To achieve that when testing,
the tests set the env var QT_TESTRUNNER_TESTING. When qt-testrunner
detects that, it prepends the command line with
sys.executable which is the currently used python interpreter.

Doing so also ensures that all sub-scripts are executed with the same
Python version.

Pick-to: 6.11 6.10 6.8
Change-Id: I4786e4015ded328077fd9cc7b443f4b4d9d69de7
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>