QUnicodeTools: port getWordBreaks() to QStringIterator
This one is a bit more complicated than the previous two, since
there's a nested loop. Solve it by using a copy of the QStringIterator
for look-ahead loop.

To see that old and new version are equivalent, observe that qsizetype
`i` and `lookahead` always pointed _onto_, while QStringIterator
always points to just _after_ the last-consumed code-unit.

Pick-to: 6.10 6.8 6.5
Change-Id: I391fafcf2418dac39b4aea3b3a3a675114233dff
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>