QUnicodeTools: port getSentenceBreaks() to QStringIterator
Like getWordBreaks(), this one is a bit more complicated than the
first two, since there's a nested loop. Solve it by using a copy of
the QStringIterator for look-ahead loop.

To see that old and new version are equivalent, observe that qsizetype
`i` and `lookahead` always pointed _onto_, while QStringIterator
always points to just _after_ the last-consumed code-unit.

Pick-to: 6.10 6.8 6.5
Change-Id: Id272b1a1597912eb611acb544b5ef0ac1d13a754
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>