Add test for focus chain when reparenting *within* the same hierarchy
For the record, focus chain is not affected by reparenting within the
window. Qt docs says:

"Tab generally moves focus in the order of widget construction"

(The exception is when a widget is moved to a separate QWidget-tree
(usually another window). This is already covered by the
tst_QWidget::focusChainOnReparent() autotest)

Change-Id: I6312b86bb81307ba11a40b6c303364056bd3d714
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>