macOS: Check entitlements directly for app sandbox detection
Checking entitlements via the code signing verification machinery of
SecStaticCodeCheckValidityWithErrors is potentially slow (may even
reach out to the network, in some cases). We were mitigating this
by spinning up a short lived thread to resolve the state before it
was needed, but pulling the sandbox state out of the entitlements
directly removes all this complexity, and is 4x faster for those
times we would end up blocking waiting for the thread to finish
(3-4ms vs 20ms).

Pick-to: 6.10 6.8
Change-Id: I069b5bb576464b6add92dc9d37fdf116f3553d0c
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>