QCoreApplication: fix memory leak in permission requests
PermissionReceiver is allocated, but never freed if
finalizePermissionRequest is not invoked. We convert it to a
shared pointer, to keep it alive via the lambda captures until the
permission request is finalized (we cannot use a unique_ptr until
move_only_function is usable). If the finalizer is not called, the
reference will be dropped and the PermissionReceiver will be destroyed.

Amends 0bd1fc006097e481d8b92ee1c5246ffa1ae9379f

Fixes: QTBUG-141432
Pick-to: 6.11 6.10
Change-Id: I996afe0be091677d2b9bedbc61dce3369a43b852
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>