QFontPrivate: properly delete the assignment operator
Before, it was implemented as a private no-op(!). This looks overly
dangerous. The typical pattern in C++98 was to declare it private,
but _not_ implement it, so one would get a linker error at least.
This, OTOH, will just compile and fail to work.

Picking to all active branches to ensure it's not inadvertently used.

Pick-to: 6.10 6.8 6.5
Change-Id: I6c2223a605665f1f0bade488fe9794d62ee335c6
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>