Remove feature conditioning from qtimezone/ tests
QTimeZone has existed even without the feature since 6.5 and its
testing then added (most of) the needed #if QT_CONIFG(timezone)
checks. I missed the CMake config when doing that, though, so its
testing was still suppressed when the feature was disabled.

This revealed that I'd missed some of the #if-ery needed in its tests,
so add what was still missing. As a drive-by, drop one #if-check from
the body of specificTransition_data(), since the whole test lies
within an outer block conditioned on the feature.

Amends commit ae6186c7e8cfdb9420b9119f5affbba7d069598d

Pick-to: 6.10 6.8 6.5
Change-Id: I0672e126cb8862563e9222fcf176ca26ac204f31
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>