QMenu: Trigger on button releases  without seeing pressed buttons before
On Wayland we can't make any assumptions about button state when
we don't have pointer focus. Crucially when opening a popup (menu in this
case), a pointer leave (focus out) event for the parent window and a
pointer enter (focus in) event for the new menu window will be sent,
and accordingly the platform backend resets the button state.
However, this means QMenu might never see a move event with pressed
buttons; and upon release, no action is triggered.

If we see a release without having seen any pressed buttons:
if the release happens over an action, activate it; otherwise
do nothing, as before (it's just release of the press that opened
the menu).

Fixes: QTBUG-124920
Pick-to: 6.10
Change-Id: Id86e81431621ce577101bf7cf45252dd1a02bfc4
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>