CMake: Fix order of Qt6*AdditionalTargetInfo.cmake file installation
Qt6*AdditionalTargetInfo.cmake files are defer installed using a
finalizer at the end of the directory scope where
qt_internal_export_additional_targets_file is called.
For tools packages, there are usually installed at the end of the root
directory scope.

qt_internal_export_additional_targets_file keeps track of which files
to install (based on an export id) in a global property.

Some tests might create a Mock Qt module which also calls the above
function and adds its own export id to the global property.

Because the mock module finalizer is called before the Tools one,
the mock finalizer will process the ids of both of the module and the
Tools package, ending up creating installation rules for the Tools
in the test directory scope.

When Qt is configured with
  -DQT_BUILD_TESTS=ON -DQT_BUILD_TESTS_BY_DEFAULT=OFF
test installation rules are skipped, which means the Tools files would
never be installed, causing configuration errors when looking up that
package later.

To avoid this situation, instead of collecting the export ids in a
global property to process all of them later, just let each finalizer
process a single id in the directory scope where the finalizer is
called.

Uncovered by d66cfa93eefa185aa15d42e9dc38cbcdbd0578cb in
qtdeclarative.

Pick-to: 6.8 6.10
Fixes: QTBUG-141466
Change-Id: Icf839835ea028df2e0079995e301693b9c3cd342
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>