a11y: Add a11y implementation for QColorWell
The QColorWell widget is used to provide/present a set of
predefined or custom colors in QColorDialog.

So far, there was no accessibility implementation
for this widget, meaning that the colors/items
were completely inaccessible on the a11y layer.

Introduce QAccessibleColorWell as new a11y
implementation for QColorWell, and
QAccessibleColorWellItem for all the
color items it provides.

As QColorWell arranges its items in columns
(i.e. color indices 0 to QColorWell::numRows() - 1
are in first column, not 0 to QColorWell::numCols() -1
in the first row), also use the same order for the
accessible children/child indices:

Send a11y focus events when the focus moves
between items in the QColorWell.

This makes both, Orca on Linux and Narrator
on Windows announce the currently focused
item when navigating through the list using
the keyboard.

For now, use "RGB <red>, <green>, <blue>"
for the accessible names of the color items
(with <red>, <green and <blue> being the value
for the color's RGB components), which is a form
that the Orca screen reader could interpret
in other contexts to speak the name of the
closest known color, see [1], section
"Speak colors as names".

[1] https://help.gnome.org/users/orca/stable/preferences_speech.html.en

Fixes: QTBUG-141571
Change-Id: I7ec64d87e6ae41070c5cb34341bbf527260cca3f
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>