WindowsStyles: Always draw QComboBoxPrivateContainer popup background
The background of QComboBox popups were not explicitly drawn. This was
no problem until QWindows11Style needed to make the popup translucent to
have rounded corners in the popup. When changing from QWindows11Style to
QWindowsVistaStyle or QWindowsStyle at runtime, the combobox popup
background became transparent due to not being explicitly filled. This
patch draws the background for popup in QWindowsVistaStyle and
QWindowsStyle explicitly.

Fixes: QTBUG-139849
Pick-to: 6.10
Change-Id: I7f312661be07eafda68f0a57fc2c32322e4da724
Reviewed-by: Christian Ehrlicher <ch.ehrlicher@gmx.de>