macOS: Use underlying NSImage for symbol images when used in system tray
There's no need to flatten the image to a QImage in this case, and
doing so loses the isTemplate state of the Image. By passing the
NSImage on, we let AppKit decide how to configure and size the
image for the menu bar, matching the look of SwiftUI's MenuBarExtra.

Note that the look of system status bar UI such as the WiFi icon
or battery icon is slightly larger than what the same icon will
look like in a MenuBarExtra, but as long as SwiftUI behaves that
way we try to match that instead of trying to compensate and align
with the few "oversized" system UI icons.

Pick-to: 6.10 6.8
Change-Id: Ia82e0bc471426345ff1b695b10c3fd463e357eb7
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>