XCB: Remove the native X11 painting engine
This is a reversal of commit 07942adb77f60738a6043665673d51fc7991233b,
which brought the older Qt 4 engine back into Qt 5. This code is
unmaintained and has only had minor bug-fixes applied to it since it was
brought back.

It doesn't remove the querying of the XRender extension that the commit
above added, because it's required by the Cursor support.

[ChangeLog][Important Changes] The experimental X11 Native Rendering
engine has been removed. Use of this engine required enabling it at
compile time with a CMake option and enabling at runtime with an
environment variable.

See https://lists.qt-project.org/pipermail/development/2025-October/046674.html

Change-Id: If8757da6a23ece83f216fffd6a27aec1f26716eb
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>