De-inline QComboBoxDelegate
Virtual functions should not be implemented inline. In particular, all
virtual functions were implemented inline (incl. the dtor), so this
class caused -Wweak-vtable, incl. all the negative effects that
entails.

As a drive-by, mark the contructor explicit, use Qt::StringLiterals,
and cache the "separator" QString that we store in the model.

Picking to all active branches to avoid conflicts going forward.

Pick-to: 6.10 6.8 6.5
Change-Id: If3e2e9c6744844fa19047d4cadb35008172cde3d
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>