De-inline QComboMenuDelegate
Virtual functions should not be implemented inline. In particular,
polymorphic types ought to have an out-of-line dtor.

In this case, not all virtual functions were implemented inline, so
this class may not have suffered from -Wweak-vtable, and all the
negative effects that entails, or at least not on all platforms.

But move the implementation into the .cpp file, anyway, incl. the
destructor, for consistency, and to reduce the number of includes in
the _p.h file.

As a drive-by, mark the contructor explicit, and the automatic
QStyleOption* objects const.

Picking to all active branches to avoid conflicts going forward.

Pick-to: 6.10 6.8 6.5
Change-Id: I321cb482586766b323a64e3be17ccf1a4984eb64
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>